Unify abstrect_sem and analyzer.go
